Refactor BalancedAllocator.Balancer to LocalShardsBalancer (#4761)
* Refactor BalancedAllocator.Balancer to LocalShardsBalancer


* Deprecate Balancer to maintain BWC


(cherry picked from commit 1d654851ff2ea50b0fbb07b7602945f3a0e4cc88)
